LayoutTests/platform/chromium/fast/events/touch/gesture/touch-gesture-scroll-iframe-propagated-expected.txt


This tests that a gesture scroll is propagated from an iframe to an outer div when the iframe has no remaining scroll offset.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


first gesture
PASS movedbox.scrollTop is 0
PASS movedbox.scrollLeft is 0
PASS wheelEventsOccurred is 0
second gesture
PASS movedbox.scrollTop is 50
PASS movedbox.scrollLeft is 0
PASS wheelEventsOccurred is 0
scroll event 0+> [object HTMLDocument]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
